Add `phylum` to system path (#11)
This is a direct copy of the same fix made in
[phylum-dev/install-phylum-latest-action](https://github.com/phylum-dev/install-phylum-latest-action/pull/13)

This isn't exactly fixing something that was broken...just making it better / more robust. That is, the action was working before this change...but that is presumed to only be the case b/c `$HOME/.local/bin` is _already_ in the `$GITHUB_PATH`.