feat: convert to a Docker container action
This change moves away from a lengthy composite action with many external action references towards a simple Docker container action. The inputs have been simplified. The permissions are well defined. The logic is completely self-contained within the `phylumio/phylum-ci` Docker image now.

BREAKING CHANGE: The inputs are different now. Tokens for GitHub and Phylum are still required, but now all the other options are encapsulated by the single `cmd` input, which is the phylum-ci command line to pass to the Docker container's entrypoint.