docs: link to external Phylum Knowledge Base FAQ
This change took the approach that replicating the FAQ(s) from the KB
FAQ page and the GitHub action documentation page is less desirable than
simplying linking to it. There is a good chance that user's of the
GitHub action will only ever see the documentation directly on GitHub
and there is a chance that documentation drift will occur if the same
information is repeated in multiple places. Additional actions taken:

* Add links to the new Phylum Project File documentation
* Shorten comment lines in the example workflow to prevent scrolling

This change is related to phylum-dev/documentation#207 but does not have
to wait for that one to be merged first.