docs: link to external Phylum Knowledge Base FAQ (#41)
This change took the approach that replicating the FAQ(s) from the KB
FAQ page and the GitHub action documentation page is less desirable than
simply linking to it. There is a good chance that users of the
GitHub action will only ever see the documentation directly on GitHub
and there is a chance that documentation drift will occur if the same
information is repeated in multiple places. Additional actions taken:

* Add links to the new Phylum Project File documentation
* Shorten comment lines in the example workflow to prevent scrolling